requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
gast==0.2.2
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
gunicorn==19.9.0
h5py==2.10.0
idna==2.8
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
numpy==1.17.3
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
Pillow==6.2.1
pipenv==2018.11.26
protobuf==3.10.0
psycopg2==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.3.1
six==1.13.0
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.25.6
virtualenv==16.7.7
virtualenv-clone==0.5.3
Werkzeug==0.16.0
whitenoise==4.1.4
wrapt==1.11.2
